:truck: Moved source code to appropiate folders
